polymarket-agent/requirements.txt
# ============================================================
# Polymarket Autonomous Trading Agent — Dependencies
# ============================================================

# Core scientific stack
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Machine learning / forecasting
torch>=2.0.0
transformers>=4.40.0

# LLM clients
openai>=1.30.0
anthropic>=0.25.0          # Optional: for Anthropic models

# Data persistence
pyarrow>=14.0.0            # For parquet caching
pyyaml>=6.0

# Polyterm dependencies (from polyterm/requirements.txt)
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
colorama>=0.4.6
tabulate>=0.9.0
click>=8.1.0

# MiroFish-inspired memory (optional: Zep Cloud)
# zep-cloud>=2.0.0          # Uncomment for full graph memory

# Utilities
python-dateutil>=2.8.2
tqdm>=4.66.0
